removing e dir gcc 4.8.0 warnings AND important fix
The important fix reverts a bad commit from imonorder.cpp, created in
abc0898874dea2cff2e55575d68ab3102a542ff1

+++ M2/Macaulay2/e/CCC.cpp
@@ -15,7 +15,7 @@ bool CCC::initialize_CCC(unsigned long prec)
   initialize_ring(0);
   declare_field();
   precision = prec;
-  _elem_size = sizeofstructtype(gmp_CC);
+  _elem_size = static_cast<int>(sizeofstructtype(gmp_CC));
   _zero_elem = new_elem();
 
   zeroV = from_int(0);
